Add support for ActionCache::UpdateActionResult().
Bazel can be configured to fall back to local builds in case remote
builds fail. They may also be configured to push such results into the
remote action cache. Leave this functionality disabled by default, as
there may be setups for which this is not desired.

At the same time, correct the error code that is returned when this
functionality is disabled. The protocol specification documents that
UNIMPLEMENTED is returned instead of PERMISSION_DENIED.