Set the Content-Length header for the remote cache storage backend.
As reported in GitHub issue #6, Bazel Remote doesn't like it if
Content-Length is not set. Setting this can be done relatively easy in
our case. For the CAS we can simply load the size from the digest. For
the AC, we never stream them in, meaning the size can also be computed.