Implementing automatic repairing of CAS inconsistencies.
I've observed cases where the CAS of a build cluster may become
corrupted. For example, S3 backend implementations that leave partially
completed uploads lying around. The Merkle blob access already allows
detecting those cases, but only causes access to the blob to abort.

Extend the Merkle blob access to send an additional delete request to
the storage backend, so that the build cluster doesn't end up in a
broken state indefinitely. If a client now needs to repeat an execute
request due to data corruption, it will properly observe that CAS
entries have become missing, so that they are re-uploaded by the client.