Add testing coverage for data corruption in MerkleBlobAccess.
While there, add proper GRPC error codes. In case of Get() operations,
we want to return codes.Internal to indicate the build setup is in a
broken state (and repairing itself). In the case of Put(), we should
return codes.InvalidArgument to indicate it's the user's fault for
trying to upload a broken blob.