Simplify size handling in the MerkleBlobAccess.
In theory, we could entirely ignore the sizeBytes argument in this case.
We could simply assume digest.SizeBytes is the one we need to use. In
case it turns out that that one is incorrect, the size checks in the
reader adapter will trap it automatically.

Still leave a log.Fatal() call in place to detect obvious misuse of this
class. It indicates that there are some issues up the stack that would
cause issues in case the MerkleBlobAccess were bypassed.