Split up MerkleBlobAccess.
For the Action Cache, we also want to validate that the digests being
passed in are somewhat valid. We don't want to allow callers to create
keys in underlying storage (e.g., Redis, S3) with arbitrary length or
characters in their name.

Add a ValidDigestRequiringBlobAccess that does the checks on the names,
while the MerkleBlobAccess will remain responsible for doing the actual
payload checks.