Massively refactor the way we deal with digests.
By using the Protobuf message for digests, there is no guarantee that
they are in a sane state. This caused us to already add the
ValidDigestRequiringBlobAccess. The problem with that approach is that
this only enforced validity of digests when accessing the CAS and AC.
Not within the layers above.

Introduce a new util.Digest type that is a combination of an instance
name and a digest Protobuf message. The constructor, NewDigest, is
written in such a way that it will disallow the creation of malformed
instances. This makes the code mostly oblivious of instance names and
hashing algorithms.