Add a function for getting trees back from the CAS.
This is an operation normally not performed by Buildbarn, as it only
emits tree objects that are downloaded by the Bazel client. I'm working
on a web frontend to explore objects in the CAS and AC. For that piece
of software we need a way of downloading trees from the CAS as well.