Add a Proof of Concept web interface for inspecting actions.
At some point we want to be able to add a status page to scheduler
processes to get insight in cluster utilization. As the scheduler does
not have any access to the CAS/AC, we should first invest some effort in
implementing a separate web frontend for at least inspecting those.

Status pages on the scheduler should be able to link to this service to
do the heavy lifting of displaying meaningful results.