Let the worker print log entries containing URLs.
The protobufs dumped in the log files by the workers contain very little
information in practice, do to them giving no access to the action,
command, input root, etc. Now that we have a somewhat decent web
frontend for browsing the CAS and AC, change this process to simply
refer to that page instead.