List output files of build actions.
First list all of the files that have been created properly (green),
sorted by type and name. Afterwards, list all of the files that are part
of the Command, but were not observed in the ActionResult. Make those
red and apply strike-through to make it obvious those did not get
created.