Let BlobAccess.Get() return the size of the blob.
In an attempt to make the Redis server usable for Action Cache entries
as well, we can no longer rely on digest.GetSizeBytes() for the "GET"
command's response header. We need to obtain the actual length from the
storage backend.

This also means that we need to extend the circular backend to keep
track of lengths in the offset store.