Add a file system abstraction to Buildbarn.
In order to deal with symbolic links in input/output files more
securely, add a file system accessor layer based on POSIX *at() calls.
These functions are available in Go using the "unix" package.
Unfortunately, that package is quite low-level.

The Directory type added by this change has member functions for the
operations that we apply to directories in the Buildbarn codebase. It
uses function signatures similar to the conventional functions that are
part of the "os" and "io/ioutil" packages.

An added advantage of this abstraction is that we crack down on the
amount of string-based path handling done by Buildbarn. It also makes
unit testing of classes that perform file system access
(LocalBuildExecutor, HardlinkingContentAddressableStorage) easier.