Rename bbb_redis_server to bbb_storage.
The downside of using the Redis protocol for Buildbarn is that the
client libraries don't support streaming of blobs. They can only be
passed in as byte arrays. This doesn't allow us to handle large blobs
properly.

The pipelining in the Redis protocol is pretty nice, but it doesn't
support adding markers/batching, meaning FindMissing() calls always end
up getting decomposed.

Because of that, experiment with simply using the bytestream.ByteStream
and remoteexecution.ContentAddressableStorage protocols. Rename the
Redis server and also let it use GRPC.