Add a BlobAccess that can access the CAS over GRPC.
Testing reveals that performance of this is on par with using Redis. The
advantage of using this over Redis is that it supports better
cancelation (through context objects), but also that it can do streaming
of larger blobs properly.

This permits us to use bbb_storage for all CAS storage.