Use dashes instead of pipes as separators in digests.
To allow multiple build actions to run concurrently, I am planning on
creating subdirectories for every build action with keys in their names.
To be specific: a key of the action digest.

Having pipes in there is going to cause lots of problems in shell
scripts with bad quoting. Resort to something safer by not using pipes.