Add support for emptying out the build directory.
When starting bbb_worker with a build directory containing garbage from
previous runs, we should make sure to clean the build directory
accordingly. Related to that, if concurrency is disabled and build
actions are executed in the root of the build directory (as done
previously), we want to make sure to empty out the build directory prior
to starting.

This change adds a series of adapters for Manager and Environment that
adds the proper logic to do this. The SingletonManager has been adjusted
to simply reject concurrent access. The CleanBuildDirectoryManager can
be wrapped around it to clean the build directory. A ConcurrentManager
can then be used to allow concurrent access to the build directory once
again.

This means that every time a worker is idle and is starting to execute
the first build action, it will attempt to clean out the build
directory.