Make the circular storage backend more robust against restarts.
The backend was already capable of auto-repairing any data corruption.
The annoying thing was, killing the process gracefully already
introduced corruption due to the write cursor being too low.

Change the overall design of the circular storage backend to no longer
write cursors to disk asynchronously. Instead, add an adapter that is
responsible for allocating more space ahead of time.

The downside of this approach is that every restart may discard some
data from storage. Restart loops will eventually cause all strage to be
invalidated. That said, that's a matter of not setting the allocation
chunk size too high.