Simply use XorShift64* to yield the right distribution.
There is no need to have dynamic memory allocation to return numbers
with the right distribution. The XorShift64* algorithm already allows us
to stretch any non-zero hash to a sequence of 2^64-1 numbers. This will
guarantee that we never reach an infinite loop, getting stuck in a cycle
where we don't return a number that corresponds to an active backend.