Tighten permissions on the build directory.
In order to make it possible at some point to add support for caching
full directory hierarchies as opposed to plain files, we should work
towards removing write permissions from directories as much as possible.

Testing reveals that restricting write permissions to just those
directories that have output files is insufficient. That tends to break
a lot of build rules, even go_proto_library(). Instead, make it so that
in addition to output directories, all parent directories (transitive)
have write permissions.

This has the advantage of removing write access from all leaves of the
tree for which we know no changes will be made. This is a powerful
property if we ever want to add support for FUSE-based builds, as it
means we can load a small number of directories initially and make them
in-memory mutable, while the rest is presented as immutable directories
on the fly.